move home components to components folder
