+++ website/src/components/home/codeShowcase.module.css
@@ -86,5 +86,3 @@
 [data-theme='dark'] .showcaseFeature:hover { background: rgba(255,255,255,0.05); border-color: rgba(255,255,255,0.15); }
 [data-theme='dark'] .showcaseFeature h4 { color: #ffffff; }
 [data-theme='dark'] .showcaseFeature p { color: rgba(255,255,255,0.7); }
-
-
