--- website/src/pages/markdown-page.md
@@ -1,8 +1,0 @@
----
-title: Markdown page example
-draft: true
----
-
-# Markdown page example
-
-You don't need React to write simple standalone pages.
